refactor: 레이어드 아키텍처를 따르도록 service 패키지명을 application 으로 변경
