+++ src/main/java/roomescape/application/ReservationAdminService.java
@@ -1,4 +1,4 @@
-package roomescape.service;
+package roomescape.application;
 
 import java.time.LocalDate;
 import org.springframework.stereotype.Service;
