+++ src/main/java/roomescape/presentation/controller/ReservationTimeController.java
@@ -15,7 +15,7 @@ import org.springframework.web.bind.annotation.RestController;
 import roomescape.dto.request.ReservationTimeRegisterDto;
 import roomescape.dto.response.AvailableReservationTimeResponseDto;
 import roomescape.dto.response.ReservationTimeResponseDto;
-import roomescape.service.ReservationTimeService;
+import roomescape.application.ReservationTimeService;
 
 @RestController
 @RequestMapping("/times")
