+++ src/main/java/roomescape/presentation/interceptor/AdminAuthInterceptor.java
@@ -7,7 +7,7 @@ import org.springframework.web.servlet.HandlerInterceptor;
 import roomescape.dto.response.MemberResponseDto;
 import roomescape.model.Role;
 import roomescape.presentation.support.CookieUtils;
-import roomescape.service.AuthService;
+import roomescape.application.AuthService;
 
 @Component
 public class AdminAuthInterceptor implements HandlerInterceptor {
