+++ src/test/java/roomescape/acceptance/AuthAcceptanceTest.java
@@ -11,7 +11,7 @@ import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.test.annotation.DirtiesContext;
 import roomescape.dto.response.MemberResponseDto;
 import roomescape.model.Role;
-import roomescape.service.JwtProvider;
+import roomescape.application.JwtProvider;
 
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.DEFINED_PORT)
 @DirtiesContext(classMode = DirtiesContext.ClassMode.BEFORE_EACH_TEST_METHOD)
