+++ src/test/java/roomescape/acceptance/MemberReservationAcceptanceTest.java
@@ -22,8 +22,8 @@ import roomescape.repository.MemberRepository;
 import roomescape.repository.ReservationRepository;
 import roomescape.repository.ReservationTimeRepository;
 import roomescape.repository.ThemeRepository;
-import roomescape.service.JwtProvider;
-import roomescape.service.ReservationService;
+import roomescape.application.JwtProvider;
+import roomescape.application.ReservationService;
 
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.DEFINED_PORT)
 @DirtiesContext(classMode = DirtiesContext.ClassMode.BEFORE_EACH_TEST_METHOD)
