+++ src/test/java/roomescape/acceptance/ReservationAcceptanceTest.java
@@ -21,7 +21,7 @@ import org.springframework.jdbc.support.KeyHolder;
 import org.springframework.test.annotation.DirtiesContext;
 import roomescape.dto.response.ReservationResponseDto;
 import roomescape.model.Role;
-import roomescape.service.JwtProvider;
+import roomescape.application.JwtProvider;
 
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.DEFINED_PORT)
 @DirtiesContext(classMode = DirtiesContext.ClassMode.BEFORE_EACH_TEST_METHOD)
