+++ src/test/java/roomescape/application/AuthServiceTest.java
@@ -1,4 +1,4 @@
-package roomescape.service;
+package roomescape.application;
 
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
