+++ src/test/java/roomescape/presentation/controller/AuthControllerTest.java
@@ -15,7 +15,7 @@ import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.
 import roomescape.dto.request.LoginRequestDto;
 import roomescape.dto.response.TokenResponseDto;
 import roomescape.presentation.support.CookieUtils;
-import roomescape.service.AuthService;
+import roomescape.application.AuthService;
 
 @WebMvcTest({AuthController.class, CookieUtils.class})
 public class AuthControllerTest {
